% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatestAccountPrice.R
\name{LatestAccountPriceUI}
\alias{LatestAccountPriceUI}
\title{最新核算价格生成生成器界面}
\usage{
LatestAccountPriceUI(
  tabTitle = "最新核算价格",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = LatestAccountPriceUI_left,
  func_right = LatestAccountPriceUI_right,
  func_bottom = LatestAccountPriceUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
最新核算价格生成生成器界面
}
\examples{
LatestAccountPriceUI()
}
