% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatestAccountPrice.R
\name{LatestAccountPriceUI_bottom}
\alias{LatestAccountPriceUI_bottom}
\title{生成区域}
\usage{
LatestAccountPriceUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
LatestAccountPriceUI_bottom()
}
