% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurPriceManagement.R
\name{PurPriceManagementUI_bottom}
\alias{PurPriceManagementUI_bottom}
\title{生成区域}
\usage{
PurPriceManagementUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
PurPriceManagementUI_bottom()
}
